backend/requirements.txt
yfinance
pandas
numpy
fastapi
uvicorn
supabase
anthropic
scikit-learn
python-dotenv